Fixed component routing and added show page
Fixed component routing to link multiple components in sequence and added a show page. Need to add more data fields to sample data in order to work on show page more. Also need to create axios call methods for add, delete, and put as well as matching UI for them