Added Login and Reg, conditonal rendering on navbar and actions columns
Added functional login and registration components. Users may be registered. Validation for ALL forms to come at a later time.

Also added conditional rendering to the Navbar for logged in users to display less options if not logged in and to NOT dispaly actions columns if the user is not logged in.