Added Search Bar Functionality
Added functionality for a search bar into the application. Created a db query to query to the SQL database that will specifically filter results based on what is typed. Currently search does NOT allow for multiple words, that will be fixed at a later date. In order to find results must type partial title.